saque cast inecesario sound player

+++ app/src/main/java/labo2018/razasypelajesonofri/utils/SoundsPlayer.java
@@ -31,7 +31,7 @@ public class SoundsPlayer {
     // COMO CUANDO NECESITAS REPRODUCIR AMBOS UNO DESPUES DEL OTRO EN EL JUEGO DE RyP JUNTOS!!
     @RequiresApi(api = Build.VERSION_CODES.JELLY_BEAN)
     private static void playSoundChain(List sounds, Context context){
-        mediaPlayers = new MediaPlayer[(int) sounds.size()];
+        mediaPlayers = new MediaPlayer[sounds.size()];
         // create MP array with respective sounds
         for (int i = 0; i < mediaPlayers.length; i++) {
             mediaPlayers[i] = MediaPlayer.create( context, (Integer) sounds.get(i));
